Refactoring of modules to clean up test warnings
