+++ src/core/mod.rs
@@ -1,6 +1,9 @@
 use clap::ValueEnum;
 use ::rand::prelude::*;
 
+#[cfg(test)]
+mod test;
+
 #[derive(ValueEnum, Debug, Clone, Default, PartialEq)]
 pub enum Target {
     Chip, // This is "chip-8" in Gulrak's opcode table
@@ -690,6 +693,7 @@ impl Chip8 {
         ((self.audio_buffer >> self.audio_oscillator as u32) & 1) as f32
     }
 
+    #[cfg(test)]
     pub fn run_frame(&mut self) {
         loop {
             self.run_inst();
