+++ src/core/test/data/mod.rs
@@ -2,7 +2,7 @@ use crate::core;
 
 pub const EMPTY_PLANE: [u128; core::HEIGHT] = [0; core::HEIGHT];
 
-pub const LOGO: &[u8] = include_bytes!("../../roms/1-chip8-logo.ch8");
+pub const LOGO: &[u8] = include_bytes!("../../../../roms/1-chip8-logo.ch8");
 pub const LOGO_FRAME_COUNT: u32 = 60;
 pub const LOGO_PLANE: [u128; core::HEIGHT] = [
     0b00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000,
@@ -71,7 +71,7 @@ pub const LOGO_PLANE: [u128; core::HEIGHT] = [
     0b00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000,
 ];
 
-pub const IBM: &[u8] = include_bytes!("../../roms/2-ibm-logo.ch8");
+pub const IBM: &[u8] = include_bytes!("../../../../roms/2-ibm-logo.ch8");
 pub const IBM_FRAME_COUNT: u32 = 60;
 pub const IBM_PLANE: [u128; core::HEIGHT] = [
     0b00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000,
@@ -140,7 +140,7 @@ pub const IBM_PLANE: [u128; core::HEIGHT] = [
     0b00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000,
 ];
 
-pub const CORAX: &[u8] = include_bytes!("../../roms/3-corax+.ch8");
+pub const CORAX: &[u8] = include_bytes!("../../../../roms/3-corax+.ch8");
 pub const CORAX_FRAME_COUNT: u32 = 180;
 pub const CORAX_PLANE: [u128; core::HEIGHT] = [
     0b00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000,
@@ -209,7 +209,7 @@ pub const CORAX_PLANE: [u128; core::HEIGHT] = [
     0b00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000,
 ];
 
-pub const FLAGS: &[u8] = include_bytes!("../../roms/4-flags.ch8");
+pub const FLAGS: &[u8] = include_bytes!("../../../../roms/4-flags.ch8");
 pub const FLAGS_FRAME_COUNT: u32 = 180;
 pub const FLAGS_PLANE: [u128; core::HEIGHT] = [
     0b10100100110011001010001100000000000000000000111000000000000000000000000000000000000000000000000000000000000000000000000000000000,
@@ -278,7 +278,7 @@ pub const FLAGS_PLANE: [u128; core::HEIGHT] = [
     0b00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000,
 ];
 
-pub const QUIRKS: &[u8] = include_bytes!("../../roms/5-quirks.ch8");
+pub const QUIRKS: &[u8] = include_bytes!("../../../../roms/5-quirks.ch8");
 pub const QUIRKS_FRAME_COUNT: u32 = 600;
 pub const QUIRKS_CHIP_PLANE: [u128; core::HEIGHT] = [
     0b00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000,
@@ -545,7 +545,7 @@ pub const QUIRKS_XO_PLANE: [u128; core::HEIGHT] = [
     0b00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000,
 ];
 
-pub const SCROLLING: &[u8] = include_bytes!("../../roms/8-scrolling.ch8");
+pub const SCROLLING: &[u8] = include_bytes!("../../../../roms/8-scrolling.ch8");
 pub const SCROLLING_FRAME_COUNT: u32 = 600;
 pub const SCROLLING_SUPER_MODERN_LOW_PLANE: [u128; core::HEIGHT] = [
     0b00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000,
